pyproject.toml
[tool.poetry]
name = "ama"
version = "0.1.0"
description = ""
authors = ["polymorphisma <shrawansunar.6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jinja2 = "3.1.2"
fastapi = "0.95.0"
uvicorn = "0.21.1"
httpx = "0.23.3"
python-multipart = "0.0.6"
faiss-cpu = "1.7.4"
langchain = "0.0.274"
textract = "1.6.5"
pandas = "2.0.3"
python-dotenv = "1.0.0"
boto3 = "1.28.52"
nltk = "3.8.1"
pyjwt = "2.8.0"
python-jose = "3.3.0"
argparse = "1.4.0"
tabulate = "0.9.0"
uuid = "^1.30"
tweepy = "^4.14.0"
gunicorn = "^22.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
